+++ Cargo.toml
@@ -4,10 +4,10 @@ members = [
     'node',
     'pallets/account-fix',
     'pallets/bitacross',
-    'pallets/bridge-related-pallets/assets-handler',
-    'pallets/bridge-related-pallets/bridge',
-    'pallets/bridge-related-pallets/bridge-transfer',
-    'pallets/bridge-related-pallets/common',
+    'pallets/bridge/assets-handler',
+    'pallets/bridge/chain-bridge',
+    'pallets/bridge/bridge-transfer',
+    'pallets/bridge/common',
     'pallets/extrinsic-filter',
     'pallets/evm-address',
     'pallets/evm-assertions',
@@ -253,11 +253,11 @@ litentry-parachain-runtime = { path = "runtime/litentry", default-features = fal
 rococo-parachain-runtime = { path = "runtime/rococo", default-features = false }
 pallet-account-fix = { path = "pallets/account-fix", default-features = false }
 pallet-asset-manager = { path = "pallets/xcm-asset-manager", default-features = false }
-pallet-assets-handler = { path = "pallets/bridge-related-pallets/assets-handler", default-features = false }
+pallet-assets-handler = { path = "pallets/bridge/assets-handler", default-features = false }
 pallet-bitacross = { path = "pallets/bitacross", default-features = false }
-pallet-bridge = { path = "pallets/bridge-related-pallets/bridge", default-features = false }
-pallet-bridge-common = { path = "pallets/bridge-related-pallets/common", default-features = false }
-pallet-bridge-transfer = { path = "pallets/bridge-related-pallets/bridge-transfer", default-features = false }
+pallet-chain-bridge = { path = "pallets/bridge/chain-bridge", default-features = false }
+pallet-bridge-common = { path = "pallets/bridge/common", default-features = false }
+pallet-bridge-transfer = { path = "pallets/bridge/bridge-transfer", default-features = false }
 pallet-extrinsic-filter = { path = "pallets/extrinsic-filter", default-features = false }
 pallet-group = { path = "pallets/group", default-features = false }
 pallet-identity-management = { path = "pallets/identity-management", default-features = false }
