--- pallets/bridge-related-pallets/assets-handler/Cargo.toml
@@ -1,56 +1,0 @@
-[package]
-authors = ['Trust Computing GmbH <info@litentry.com>']
-description = 'Pallet for Assets handler of token bridge'
-edition = '2021'
-homepage = 'https://litentry.com/'
-license = 'GPL-3.0'
-name = 'pallet-assets-handler'
-repository = 'https://github.com/litentry/litentry-parachain'
-version = '0.1.0'
-
-[dependencies]
-hex-literal = { workspace = true }
-parity-scale-codec = { workspace = true }
-scale-info = { workspace = true }
-
-frame-support = { workspace = true }
-frame-system = { workspace = true }
-pallet-assets = { workspace = true }
-pallet-balances = { workspace = true }
-pallet-bridge = { workspace = true }
-pallet-bridge-common = { workspace = true }
-pallet-parachain-staking = { workspace = true }
-sp-core = { workspace = true }
-sp-io = { workspace = true }
-sp-runtime = { workspace = true }
-sp-std = { workspace = true }
-
-[dev-dependencies]
-pallet-bridge-transfer = { workspace = true, features = ["std"] }
-
-[features]
-default = ["std"]
-runtime-benchmarks = [
-    "frame-support/runtime-benchmarks",
-    "frame-system/runtime-benchmarks",
-    "pallet-assets/runtime-benchmarks",
-    "pallet-bridge/runtime-benchmarks",
-    "pallet-bridge-common/runtime-benchmarks",
-    "pallet-parachain-staking/runtime-benchmarks",
-]
-std = [
-    "parity-scale-codec/std",
-    "scale-info/std",
-    "sp-std/std",
-    "sp-runtime/std",
-    "sp-io/std",
-    "sp-core/std",
-    "frame-support/std",
-    "frame-system/std",
-    "pallet-assets/std",
-    "pallet-balances/std",
-    "pallet-bridge/std",
-    "pallet-bridge-common/std",
-    "pallet-parachain-staking/std",
-]
-try-runtime = ["frame-support/try-runtime"]
