+++ pallets/bridge/assets-handler/src/tests.rs
@@ -34,7 +34,10 @@ fn make_transfer_proposal(to: u64, amount: u64) -> RuntimeCall {
 
 #[test]
 fn constant_equality() {
-	let r_id = pallet_bridge::derive_resource_id(1, &pallet_bridge::hashing::blake2_128(b"LIT"));
+	let r_id = pallet_chain_bridge::derive_resource_id(
+		1,
+		&pallet_chain_bridge::hashing::blake2_128(b"LIT"),
+	);
 	let encoded: [u8; 32] =
 		hex!("0000000000000000000000000000000a21dfe87028f214dd976be8479f5af001");
 	assert_eq!(r_id, encoded);
@@ -42,7 +45,7 @@ fn constant_equality() {
 
 #[test]
 fn transfer() {
-	let dest_bridge_id: pallet_bridge::BridgeChainId = 0;
+	let dest_bridge_id: pallet_chain_bridge::BridgeChainId = 0;
 	let resource_id = NativeTokenResourceId::get();
 	let native_token_asset_info: AssetInfo<
 		<Test as pallet_assets::Config>::AssetId,
@@ -77,7 +80,7 @@ fn transfer() {
 
 #[test]
 fn transfer_assets() {
-	let dest_bridge_id: pallet_bridge::BridgeChainId = 0;
+	let dest_bridge_id: pallet_chain_bridge::BridgeChainId = 0;
 	let resource_id = NativeTokenResourceId::get();
 	let native_token_asset_info: AssetInfo<
 		<Test as pallet_assets::Config>::AssetId,
@@ -117,7 +120,7 @@ fn transfer_assets() {
 					amount: 100,
 					fee: 10,
 				}),
-				RuntimeEvent::Bridge(pallet_bridge::Event::FungibleTransfer(
+				RuntimeEvent::Bridge(pallet_chain_bridge::Event::FungibleTransfer(
 					dest_bridge_id,
 					1,
 					resource_id,
@@ -131,7 +134,7 @@ fn transfer_assets() {
 
 #[test]
 fn mint_overflow() {
-	let dest_bridge_id: pallet_bridge::BridgeChainId = 0;
+	let dest_bridge_id: pallet_chain_bridge::BridgeChainId = 0;
 	let resource_id = NativeTokenResourceId::get();
 	let native_token_asset_info: AssetInfo<
 		<Test as pallet_assets::Config>::AssetId,
@@ -160,9 +163,9 @@ fn mint_overflow() {
 fn transfer_to_regular_account() {
 	new_test_ext().execute_with(|| {
 		let dest_chain = 0;
-		let asset = pallet_bridge::derive_resource_id(
+		let asset = pallet_chain_bridge::derive_resource_id(
 			dest_chain,
-			&pallet_bridge::hashing::blake2_128(b"an asset"),
+			&pallet_chain_bridge::hashing::blake2_128(b"an asset"),
 		);
 		let amount: u64 = 100;
 
@@ -180,7 +183,7 @@ fn transfer_to_regular_account() {
 
 #[test]
 fn create_successful_transfer_proposal() {
-	let src_id: pallet_bridge::BridgeChainId = 0;
+	let src_id: pallet_chain_bridge::BridgeChainId = 0;
 	let r_id = NativeTokenResourceId::get();
 	let native_token_asset_info: AssetInfo<
 		<Test as pallet_assets::Config>::AssetId,
@@ -200,10 +203,10 @@ fn create_successful_transfer_proposal() {
 			Box::new(proposal.clone())
 		));
 		let prop = Bridge::votes(src_id, (prop_id, proposal.clone())).unwrap();
-		let expected = pallet_bridge::ProposalVotes {
+		let expected = pallet_chain_bridge::ProposalVotes {
 			votes_for: vec![RELAYER_A],
 			votes_against: vec![],
-			status: pallet_bridge::ProposalStatus::Initiated,
+			status: pallet_chain_bridge::ProposalStatus::Initiated,
 			expiry: ProposalLifetime::get() + 1,
 		};
 		assert_eq!(prop, expected);
@@ -217,10 +220,10 @@ fn create_successful_transfer_proposal() {
 			Box::new(proposal.clone())
 		));
 		let prop = Bridge::votes(src_id, (prop_id, proposal.clone())).unwrap();
-		let expected = pallet_bridge::ProposalVotes {
+		let expected = pallet_chain_bridge::ProposalVotes {
 			votes_for: vec![RELAYER_A],
 			votes_against: vec![RELAYER_B],
-			status: pallet_bridge::ProposalStatus::Initiated,
+			status: pallet_chain_bridge::ProposalStatus::Initiated,
 			expiry: ProposalLifetime::get() + 1,
 		};
 		assert_eq!(prop, expected);
@@ -234,10 +237,10 @@ fn create_successful_transfer_proposal() {
 			Box::new(proposal.clone())
 		));
 		let prop = Bridge::votes(src_id, (prop_id, proposal)).unwrap();
-		let expected = pallet_bridge::ProposalVotes {
+		let expected = pallet_chain_bridge::ProposalVotes {
 			votes_for: vec![RELAYER_A, RELAYER_C],
 			votes_against: vec![RELAYER_B],
-			status: pallet_bridge::ProposalStatus::Approved,
+			status: pallet_chain_bridge::ProposalStatus::Approved,
 			expiry: ProposalLifetime::get() + 1,
 		};
 		assert_eq!(prop, expected);
@@ -245,17 +248,19 @@ fn create_successful_transfer_proposal() {
 		assert_eq!(Balances::free_balance(RELAYER_A), ENDOWED_BALANCE + 10);
 
 		assert_events(vec![
-			RuntimeEvent::Bridge(pallet_bridge::Event::VoteFor(src_id, prop_id, RELAYER_A)),
-			RuntimeEvent::Bridge(pallet_bridge::Event::VoteAgainst(src_id, prop_id, RELAYER_B)),
-			RuntimeEvent::Bridge(pallet_bridge::Event::VoteFor(src_id, prop_id, RELAYER_C)),
-			RuntimeEvent::Bridge(pallet_bridge::Event::ProposalApproved(src_id, prop_id)),
+			RuntimeEvent::Bridge(pallet_chain_bridge::Event::VoteFor(src_id, prop_id, RELAYER_A)),
+			RuntimeEvent::Bridge(pallet_chain_bridge::Event::VoteAgainst(
+				src_id, prop_id, RELAYER_B,
+			)),
+			RuntimeEvent::Bridge(pallet_chain_bridge::Event::VoteFor(src_id, prop_id, RELAYER_C)),
+			RuntimeEvent::Bridge(pallet_chain_bridge::Event::ProposalApproved(src_id, prop_id)),
 			RuntimeEvent::AssetsHandler(Event::TokenBridgeIn {
 				asset_id: None,
 				to: RELAYER_A,
 				amount: 10,
 			}),
 			RuntimeEvent::Balances(pallet_balances::Event::Minted { who: RELAYER_A, amount: 10 }),
-			RuntimeEvent::Bridge(pallet_bridge::Event::ProposalSucceeded(src_id, prop_id)),
+			RuntimeEvent::Bridge(pallet_chain_bridge::Event::ProposalSucceeded(src_id, prop_id)),
 		]);
 	})
 }
@@ -360,7 +365,7 @@ fn test_external_balances_adjusted() {
 			resource_id,
 			native_token_asset_info
 		));
-		assert_ok!(pallet_bridge::Pallet::<Test>::whitelist_chain(
+		assert_ok!(pallet_chain_bridge::Pallet::<Test>::whitelist_chain(
 			RuntimeOrigin::root(),
 			dest_chain
 		));
