+++ pallets/bridge/bridge-transfer/Cargo.toml
@@ -2,16 +2,17 @@
 authors = ['Trust Computing GmbH <info@litentry.com>']
 edition = '2021'
 homepage = 'https://litentry.com'
-name = 'pallet-bridge'
+name = 'pallet-bridge-transfer'
 repository = 'https://github.com/litentry/litentry-pallets'
 version = '0.1.0'
 
 [dependencies]
+hex-literal = { workspace = true }
 parity-scale-codec = { workspace = true }
 scale-info = { workspace = true }
 
+sp-arithmetic = { workspace = true }
 sp-core = { workspace = true }
-sp-core-hashing = { workspace = true }
 sp-io = { workspace = true }
 sp-runtime = { workspace = true }
 sp-std = { workspace = true }
@@ -19,7 +20,15 @@ sp-std = { workspace = true }
 frame-benchmarking = { workspace = true, optional = true }
 frame-support = { workspace = true }
 frame-system = { workspace = true }
-pallet-balances = { workspace = true }
+pallet-assets = { workspace = true }
+
+pallet-bridge-common = { workspace = true }
+pallet-chain-bridge = { workspace = true }
+
+[dev-dependencies]
+pallet-assets-handler = { workspace = true, features = ["std"] }
+pallet-balances = { workspace = true, features = ["std"] }
+pallet-timestamp = { workspace = true, features = ["std"] }
 
 [features]
 default = ["std"]
@@ -27,6 +36,9 @@ runtime-benchmarks = [
     "frame-benchmarking/runtime-benchmarks",
     "frame-support/runtime-benchmarks",
     "frame-system/runtime-benchmarks",
+    "pallet-assets/runtime-benchmarks",
+    "pallet-chain-bridge/runtime-benchmarks",
+    "pallet-bridge-common/runtime-benchmarks",
 ]
 std = [
     "parity-scale-codec/std",
@@ -35,10 +47,12 @@ std = [
     "sp-runtime/std",
     "sp-io/std",
     "sp-core/std",
-    "sp-core-hashing/std",
+    "sp-arithmetic/std",
     "frame-support/std",
     "frame-system/std",
     "frame-benchmarking?/std",
-    "pallet-balances/std",
+    "pallet-assets/std",
+    "pallet-chain-bridge/std",
+    "pallet-bridge-common/std",
 ]
 try-runtime = ["frame-support/try-runtime"]
