+++ pallets/bridge/chain-bridge/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+authors = ['Trust Computing GmbH <info@litentry.com>']
+edition = '2021'
+homepage = 'https://litentry.com'
+name = 'pallet-chain-bridge'
+repository = 'https://github.com/litentry/litentry-pallets'
+version = '0.1.0'
+
+[dependencies]
+parity-scale-codec = { workspace = true }
+scale-info = { workspace = true }
+
+sp-core = { workspace = true }
+sp-core-hashing = { workspace = true }
+sp-io = { workspace = true }
+sp-runtime = { workspace = true }
+sp-std = { workspace = true }
+
+frame-benchmarking = { workspace = true, optional = true }
+frame-support = { workspace = true }
+frame-system = { workspace = true }
+pallet-balances = { workspace = true }
+
+[features]
+default = ["std"]
+runtime-benchmarks = [
+    "frame-benchmarking/runtime-benchmarks",
+    "frame-support/runtime-benchmarks",
+    "frame-system/runtime-benchmarks",
+]
+std = [
+    "parity-scale-codec/std",
+    "scale-info/std",
+    "sp-std/std",
+    "sp-runtime/std",
+    "sp-io/std",
+    "sp-core/std",
+    "sp-core-hashing/std",
+    "frame-support/std",
+    "frame-system/std",
+    "frame-benchmarking?/std",
+    "pallet-balances/std",
+]
+try-runtime = ["frame-support/try-runtime"]
