+++ pallets/bridge/chain-bridge/src/hashing.rs
