+++ runtime/litentry/src/lib.rs
@@ -870,14 +870,14 @@ parameter_types! {
 	pub TreasuryAccount: AccountId = TreasuryPalletId::get().into_account_truncating();
 }
 
-impl pallet_bridge::Config for Runtime {
+impl pallet_chain_bridge::Config for Runtime {
 	type RuntimeEvent = RuntimeEvent;
 	type BridgeCommitteeOrigin = EnsureRootOrHalfCouncil;
 	type Proposal = RuntimeCall;
 	type BridgeChainId = BridgeChainId;
 	type Balance = Balance;
 	type ProposalLifetime = ProposalLifetime;
-	type WeightInfo = weights::pallet_bridge::WeightInfo<Runtime>;
+	type WeightInfo = weights::pallet_chain_bridge::WeightInfo<Runtime>;
 }
 
 parameter_types! {
@@ -907,7 +907,7 @@ impl SortedMembers<AccountId> for TransferAssetsAnyone {
 
 impl pallet_bridge_transfer::Config for Runtime {
 	type BridgeHandler = AssetsHandler;
-	type BridgeOrigin = pallet_bridge::EnsureBridge<Runtime>;
+	type BridgeOrigin = pallet_chain_bridge::EnsureBridge<Runtime>;
 	type TransferAssetsMembers = TransferAssetsAnyone;
 	type WeightInfo = weights::pallet_bridge_transfer::WeightInfo<Runtime>;
 }
@@ -1192,7 +1192,7 @@ construct_runtime! {
 		Assets: pallet_assets = 56,
 
 		// Litentry pallets
-		ChainBridge: pallet_bridge = 60,
+		ChainBridge: pallet_chain_bridge= 60,
 		BridgeTransfer: pallet_bridge_transfer = 61,
 		ExtrinsicFilter: pallet_extrinsic_filter = 63,
 		AssetManager: pallet_asset_manager = 64,
@@ -1328,7 +1328,7 @@ mod benches {
 		[cumulus_pallet_xcmp_queue, XcmpQueue]
 		[pallet_identity_management, IdentityManagement]
 		[pallet_vc_management, VCManagement]
-		[pallet_bridge,ChainBridge]
+		[pallet_chain_bridge,ChainBridge]
 		[pallet_bridge_transfer,BridgeTransfer]
 		[pallet_teebag, Teebag]
 	);
