+++ runtime/litentry/src/weights/mod.rs
@@ -20,8 +20,8 @@ pub mod cumulus_pallet_xcmp_queue;
 // pub mod frame_system;
 pub mod pallet_asset_manager;
 pub mod pallet_balances;
-pub mod pallet_bridge;
 pub mod pallet_bridge_transfer;
+pub mod pallet_chain_bridge;
 pub mod pallet_collective;
 pub mod pallet_democracy;
 pub mod pallet_evm;
