+++ runtime/rococo/Cargo.toml
@@ -81,8 +81,8 @@ pallet-asset-manager = { workspace = true }
 pallet-assets = { workspace = true }
 pallet-assets-handler = { workspace = true }
 pallet-bitacross = { workspace = true }
-pallet-bridge = { workspace = true }
 pallet-bridge-transfer = { workspace = true }
+pallet-chain-bridge = { workspace = true }
 pallet-evm-assertions = { workspace = true }
 pallet-extrinsic-filter = { workspace = true }
 pallet-group = { workspace = true }
@@ -161,7 +161,7 @@ runtime-benchmarks = [
     "runtime-common/runtime-benchmarks",
     "sp-runtime/runtime-benchmarks",
     "xcm-builder/runtime-benchmarks",
-    "pallet-bridge/runtime-benchmarks",
+    "pallet-chain-bridge/runtime-benchmarks",
     "pallet-bridge-transfer/runtime-benchmarks",
     "pallet-extrinsic-filter/runtime-benchmarks",
     "cumulus-pallet-parachain-system/runtime-benchmarks",
@@ -259,7 +259,7 @@ std = [
     'runtime-common/std',
     "pallet-account-fix/std",
     "pallet-bitacross/std",
-    "pallet-bridge/std",
+    "pallet-chain-bridge/std",
     "pallet-bridge-transfer/std",
     "pallet-evm-assertions/std",
     "pallet-extrinsic-filter/std",
@@ -295,7 +295,7 @@ try-runtime = [
     "pallet-balances/try-runtime",
     "pallet-bitacross/try-runtime",
     "pallet-bounties/try-runtime",
-    "pallet-bridge/try-runtime",
+    "pallet-chain-bridge/try-runtime",
     "pallet-bridge-transfer/try-runtime",
     "pallet-collective/try-runtime",
     "pallet-democracy/try-runtime",
